feat: use preferred origin (part 2)
It selects the preferred origin of each event prepared as  hypoDD
input. The default behavior was to assume that each event only has
one origin. Events with only one origin are not affected by this
commit. The method _create_plots() had to be slightly modified
so as to not break the plot creation.

Closes issue #22